requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.1.0
torchvision==0.16.0
Pillow
accelerate
transformers
diffusers
# VITON-HD requirements
opencv-python
torchgeometry
# The IMAGDressing repository is not pip-installable directly (it lacks setup.py/pyproject).
# We install it during the Docker build by cloning and making it installable. Keep this commented
# if you want to install locally via git in the future, transform into a proper package.
# git+https://github.com/muzishen/IMAGDressing.git
